tracking/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.11.8
cupy-cuda100==7.1.1
cycler==0.10.0
decorator==4.4.2
easydict==1.9
fastrlock==0.5
gast==0.4.0
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
imageio==2.9.0
importlib-metadata==2.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-contrib-python==4.4.0.46
opencv-python==4.1.2.30
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==1.14.0
tensorboardX==2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tifffile==2020.10.1
torch==1.2.0
torchvision==0.4.1
tqdm==4.49.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
